Get rid of EBNF syntax in comments on grammar spec/rules
Now we use D's Pegged formalism:
<https://github.com/dlang-community/Pegged/wiki/PEG-Basics>.

There's only one extension used to this formalism: the special undefined
rule Right1. Right1 can only be used in rules in the form:

Rule <- Rule AnotherRule Right1 Rule

That is to say it that it can only be used in rules with left and right
recursion. It's used to change the associativity of binary operators
being defined by some rule. If Right1 is used, associativity will be
left-to-right. If unused, associativity will be right-to-left.

It only affects the rule that immediately follows, and that must be a
use of direct-recursion of the rule being defined. Parentheses are
unnecessary, but it makes the syntax easier to use, so that's what
you'll most likely see in the comments/spec.