Memoize recursive invocations of PEG rules
This saves a shitload of time (at the cost of a shitload of RAM). On my
system (AMD Ryzen 9 5950X), processing a random 1.2K file w/o
memoization:

0.93s user 0.00s system 99% cpu 0.936 total

When memoization is done:

0.00s user 0.00s system 91% cpu 0.002 total

I've done the above tests with the code modified to make use of mmap to
read the file from a tmpfs so IO wouldn't interfere with parse time.

When the file gets larger, the parser w/o memoization takes such
unreasonable time that it just becomes useless.

The memoized PEG parser can still be slow (e.g. it takes 20 seconds to
process a random 7.4M file with 20 levels of recursive expressions on my
system), but now it's good enough to process the size of data that I'm
expecting it to be used for. For the outliers, it'll be very slow, but
at least it'll still finish (unless the input size is really
unreasonable).