Allow recursion_context::enter() to use "any" function
Now one can always use the pattern recur.enter<Fn>(reader) even if Fn
isn't declared in the memoization table. The goal is to further unify
the code patterns so that eventually it becomes possible to do what
Guido did at
https://medium.com/@gvanrossum_83706/generating-a-peg-parser-520057d642a9

The parser generator I have in mind would allow users to provide custom
code to parse some of the rules in the grammar. The parser generator
shouldn't have to care whether one of such rules is going to participate
in left-recursions or memoization. The parser generator should always be
able to call them using the same pattern. Hence this commit.

This change also makes it easier to refactor grammars later w/o having
to change call sites just because the grammar changed to allow new
recursions. It should be more error-proof.