Fix AST printing
Passing a shared_ptr as a BoolExpr parameter was triggering the
constructor for NegExpr which takes a single shared_ptr as argument.

As a safety measure to avoid this mistake again, we could turn NegExpr
constructor as an explicit constructor. However the constructors will
likely change anyways to take line and column directly and avoid the
visitation in make_expr() so I think it's not worth to pursue such
safety measures now.