Remove unneeded logic in reader
Previously it'd be possible for reader to point to
before-the-first-token. When such scenario was possible, a call to
next() would move cursor leaving begin untouched. For a reader pointing
to before-the-first-token and a reader pointing to the first token,
begin would compare equals, and a comparison to cursor was then
necessary to disambiguate.

Now there's a call to next() within the reader constructor and it's no
longer possible to have a reader object pointing to
before-the-first-token.

Now it's no longer necessary to compare against cursor to disambiguate.